volksmarkt-backend-main/requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
altgraph==0.17.3
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asgiref==3.5.2
asttokens==2.2.1
async-generator==1.10
async-timeout==4.0.2
asyncio==3.4.3
attrs==22.2.0
backcall==0.2.0
certifi==2022.12.7
cffi==1.15.1
chardet==3.0.4
charset-normalizer==2.1.1
click==8.1.3
comm==0.1.2
coreapi==2.3.3
coreschema==0.0.4
cryptography==39.0.0
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.6
Django==4.1.1
django-cors-headers==3.13.0
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
et-xmlfile==1.1.0
executing==1.2.0
fake-useragent==0.1.10
fastjsonschema==2.16.2
filelock==3.9.0
fqdn==1.5.1
frozenlist==1.3.3
future==0.18.2
h11==0.14.0
html5lib==1.0.1
idna==3.4
isoduration==20.11.0
itsdangerous==2.1.2
itypes==1.2.0
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jsonpointer==2.3
jsonschema==4.17.3
MarkupSafe==2.1.1
matplotlib-inline==0.1.6
mistune==2.0.5
multidict==6.0.3
nbclassic==0.5.1
nbclient==0.7.2
nbconvert==7.2.9
nbformat==5.7.3
nest-asyncio==1.5.6
notebook==6.5.2
notebook_shim==0.2.2
numpy==1.23.4
oauthlib==3.2.2
outcome==1.2.0
packaging==23.0
pandas==1.5.1
pandocfilters==1.5.0
parso==0.8.3
pefile==2022.5.30
pickleshare==0.7.5
Pillow==9.3.0
pipenv==2022.12.19
platformdirs==2.6.2
prettytable==0.7.2
prometheus-client==0.16.0
prompt-toolkit==3.0.36
psutil==5.9.4
pure-eval==0.2.2
pycparser==2.21
pyinstaller==5.6.2
pyinstaller-hooks-contrib==2022.13
PyJWT==2.6.0
pymongo==3.6.1
pyOpenSSL==23.0.0
pyrsistent==0.19.3
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.21.0
python-json-logger==2.0.4
python3-openid==3.2.0
pytz==2022.6
PyYAML==6.0
pyzmq==25.0.0
qtconsole==5.4.0
QtPy==2.3.0
requests==2.28.2
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
six==1.16.0
sniffio==1.3.0
social-auth-app-django==4.0.0
social-auth-core==4.4.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
stack-data==0.6.2
termcolor==2.2.0
terminado==0.17.1
threadpoolctl==3.1.0
tinycss2==1.2.1
tornado==6.2
tqdm==4.64.1
traitlets==5.9.0
trio==0.22.0
trio-websocket==0.9.2
tzdata==2022.7
uri-template==1.2.0
uritemplate==4.1.1
urllib3==1.26.14
urllib3-secure-extra==0.1.0
virtualenv==20.17.1
virtualenv-clone==0.5.7
wcwidth==0.2.6
webcolors==1.12
webencodings==0.5.1
websocket-client==1.5.1
Werkzeug==2.2.2
widgetsnbextension==4.0.5
wrapt==1.11.2
wsproto==1.2.0
yarl==1.8.2
